+++ PES-IDS/c_jac_1_2_1_4_purify.h
@@ -2,6 +2,6 @@
 #define C_JAC_1_2_1_4_PURIFY_H
 
 #include <Eigen/Dense>
-extern "C" void evpoly_jac_1_2_1_4_purify(Eigen::Ref<Eigen::MatrixXf> jac, float* y);
+extern "C" void evpoly_jac_1_2_1_4_purify(Eigen::Ref<Eigen::MatrixXd> jac, double* y);
 
 #endif
